+++ module-domain/src/main/java/com/whiskey/domain/payment/dto/PaymentCommand.java
@@ -1,8 +1,10 @@
 package com.whiskey.domain.payment.dto;
 
+import java.math.BigDecimal;
+
 public record PaymentCommand(
     Long memberId,
     Long orderId,
-    Long amount,
+    BigDecimal amount,
     String description
 ) {}
